pyproject.toml
[tool.poetry]
name = "photoflow"
version = "0.1.0"
description = ""
authors = ["Artem Yurkov <yurkov.artyom21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "3.2.16"
python-dotenv = "^1.0.1"
django-bootstrap-v5 = "^1.0.11"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
